+++ src/main/java/step5/domain/Car.java
@@ -1,4 +1,4 @@
-package step5;
+package step5.domain;
 
 import java.util.Objects;
